// src/modules/user/enums/user-zodiac-sign.enum.ts
export enum USER_ZODIAC_SIGN {
  ARIES = 'Áries',
  TAURUS = 'Touro',
  GEMINI = 'Gêmeos',
  CANCER = 'Câncer',
  LEO = 'Leão',
  VIRGO = 'Virgem',
  LIBRA = 'Libra',
  SCORPIO = 'Escorpião',
  SAGITTARIUS = 'Sagitário',
  CAPRICORN = 'Capricórnio',
  AQUARIUS = 'Aquário',
  PISCES = 'Peixes',
}